pyproject.toml
[tool.poetry]
name = "icomplete"
version = "0.1.0"
description = ""
authors = ["Giulio Vogric <giulio.vogric@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["iComplete", "pools", "migrations"]
venvPath = "."
venv = ".venv"
